Re-arranged + Docker updates.
