+++ Dockerfile
@@ -1,8 +1,12 @@
+# Dockerfile
 FROM node:20-alpine
 
 RUN corepack enable && corepack prepare pnpm@latest --activate
 
 WORKDIR /app
+
+# Install bash for our scripts
+RUN apk add --no-cache bash
 
 COPY package.json pnpm-lock.yaml ./
 
@@ -10,8 +14,13 @@ RUN pnpm install
 
 COPY . .
 
-ENV NODE_ENV=production
+ENV NODE_ENV=production \
+    POLLING_INTERVAL=180
+
+# Make our entry point script executable
+RUN chmod +x docker-entrypoint.sh
 
 EXPOSE 3000
 
+ENTRYPOINT ["/app/docker-entrypoint.sh"]
 CMD ["pnpm", "start"]
