+++ src/constants/others.js
@@ -1,0 +1,9 @@
+const PRICE_CACHE_KEY = "mina:latest_price";
+const POLLING_INTERVAL = 3 * 60 * 1000; // Can be changed. In ms.
+const MULTIPLICATION_FACTOR = 10;
+
+module.exports = {
+  PRICE_CACHE_KEY,
+  POLLING_INTERVAL,
+  MULTIPLICATION_FACTOR,
+};
