+++ src/utils/clients/redis.js
@@ -14,8 +14,4 @@ redis.on("error", (err) => {
   console.error("Redis Error:", err);
 });
 
-redis.on("connect", () => {
-  console.log("Successfully connected to Redis");
-});
-
 module.exports = { redis };
